backend/services/organization-service/requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
httpx>=0.24.0
python-multipart>=0.0.6
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.1
python-dotenv>=1.0.0

# GraphQL and Federation dependencies
strawberry-graphql[fastapi]>=0.200.0
starlette>=0.27.0

# Google Cloud API dependencies
google-auth>=2.17.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0

# Supabase
supabase>=2.0.0

# Testing
pytest>=7.3.1
pytest-asyncio>=0.21.0
